refactor: 메시지 출력 양식 및 반환타입 변경

+++ src/main/java/kr/co/programmers/springbootbasic/io/Input.java
@@ -5,5 +5,5 @@ import kr.co.programmers.springbootbasic.voucher.VoucherType;
 public interface Input {
     MenuCommand readMenuCommand();
     VoucherType readVoucherType();
-    int readAmount();
+    long readAmount();
 }
